Auto merged
2002/09/29 21:53:51+03:00 staelin.co.il!staelin
- change return type for bytes() to uint64, and switch from atoi() to
  sscanf() to convert from string to integer

(Logical change 1.200)

+++ trunk/lmbench3/src/timing.h
@@ -10,7 +10,7 @@ double	Delta(void);
 double	Now(void);
 void	adjust(int usec);
 void	bandwidth(uint64 bytes, uint64 times, int verbose);
-int	bytes(char *s);
+uint64	bytes(char *s);
 void	context(uint64 xfers);
 uint64	delta(void);
 int	get_enough(int);
